// generated/src/app/api-models/procedure-price.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProcedurePrice {
  procedureCode: string;
  seqProcPrice: number;
  priceSchedule: string;
  pricingRegion: string;
  effectiveDate: Date;
  termDate: Date;
  allowedAmt: number;
  pctOfBilled: number;
  withholdPct: number;
  contractOverride: string;
  procedureHold: string;
  holdDate: Date;
  perDiemFlag: string;
  userDefined1: string;
  userDefined2: string;
  userDefined3: string;
  userDefined4: string;
  userDefined5: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  modifierCode: string;
  geoZipRegion: string;
}